Updated system labels
Updated and simplified label on the LivelyGig alias.
Updated some of the labels for ProjectPost, TalentPost to match the
current UI design.
Removed AgentTypeSpec, since these labels have to be simple and are
mentioned in the Overview Of LivelyGig Protocols --
https://docs.google.com/document/d/1471uLc58tlO3NGxtsKw_UuzzofbMTDhdFS5uJxx2MGA/edit#.
Add connection date, so a user can sort by recent connections.